# assets/swagger/specs/v1/spec.json
{"swagger":"2.0","info":{"description":"This is Swagger 2.0 compliant client for AnyPresence Node generated API","version":"v1","title":"HTTP REST API","termsOfService":"http://www.anypresence.com/terms/","contact":{"name":"info@anypresence.com"}},"basePath":"/api","schemes":["http"],"paths":{"/v1/publishes":{"get":{"tags":["publishes"],"summary":"Get collection of all resources","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"scope","in":"query","required":false,"type":"string","description":"Use one of the defined query scopes:  all, exact_match, count, count_exact_match"},{"name":"query[id]","in":"query","required":false,"type":"integer"},{"name":"query[channel]","in":"query","required":false,"type":"string"},{"name":"query[message]","in":"query","required":false,"type":"string"},{"name":"query[pubkey]","in":"query","required":false,"type":"string"},{"name":"query[sig]","in":"query","required":false,"type":"string"},{"name":"query[subkey]","in":"query","required":false,"type":"string"},{"name":"query[timestamp]","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Resource representation successfully returned"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}},"post":{"tags":["publishes"],"summary":"Adds a new resource using representation provided in body","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Representation of the resource that is going to be updated","required":false,"schema":{"$ref":"#/definitions/publishes"}}],"responses":{"201":{"description":"Resource was successfully created"},"422":{"description":"Resource could not be created due to business validations were not achieved"},"500":{"description":"Error trying to process request"}}}},"/v1/publishes/{id}":{"get":{"tags":["publishes"],"summary":"Get an existing representation of the given resource","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"}],"responses":{"200":{"description":"Resource representation successfully returned"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}},"put":{"tags":["publishes"],"summary":"Update an existing resource using representation provided in request body","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"},{"in":"body","name":"body","description":"Representation of the resource that is going to be updated","required":false,"schema":{"$ref":"#/definitions/publishes"}}],"responses":{"204":{"description":"Resource was successfully updated"},"404":{"description":"Requested resource does not exist"},"422":{"description":"Resource could not be created due to business validations were not achieved"},"500":{"description":"Error trying to process request"}}},"delete":{"tags":["publishes"],"summary":"Deletes an existing representation of a given resource","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"}],"responses":{"204":{"description":"Resource was successfully updated"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}}},"/push_notifications/channel":{"get":{"tags":["channel"],"summary":"Get collection of all resources","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"scope","in":"query","required":false,"type":"string","description":"Use one of the defined query scopes:  all, count_exact_match, count, exact_match"},{"name":"query[name]","in":"query","required":false,"type":"string"},{"name":"query[devices]","in":"query","required":false,"type":"string"},{"name":"query[messages]","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Resource representation successfully returned"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}},"post":{"tags":["channel"],"summary":"Adds a new resource using representation provided in body","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Representation of the resource that is going to be updated","required":false,"schema":{"$ref":"#/definitions/channel"}}],"responses":{"201":{"description":"Resource was successfully created"},"422":{"description":"Resource could not be created due to business validations were not achieved"},"500":{"description":"Error trying to process request"}}}},"/push_notifications/channel/{id}":{"get":{"tags":["channel"],"summary":"Get an existing representation of the given resource","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"}],"responses":{"200":{"description":"Resource representation successfully returned"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}},"put":{"tags":["channel"],"summary":"Update an existing resource using representation provided in request body","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"},{"in":"body","name":"body","description":"Representation of the resource that is going to be updated","required":false,"schema":{"$ref":"#/definitions/channel"}}],"responses":{"204":{"description":"Resource was successfully updated"},"404":{"description":"Requested resource does not exist"},"422":{"description":"Resource could not be created due to business validations were not achieved"},"500":{"description":"Error trying to process request"}}},"delete":{"tags":["channel"],"summary":"Deletes an existing representation of a given resource","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"}],"responses":{"204":{"description":"Resource was successfully updated"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}}},"/push_notifications/message":{"get":{"tags":["message"],"summary":"Get collection of all resources","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"scope","in":"query","required":false,"type":"string","description":"Use one of the defined query scopes:  all, count_exact_match, count, exact_match"},{"name":"query[time]","in":"query","required":false,"type":"string"},{"name":"query[receiver]","in":"query","required":false,"type":"string"},{"name":"query[is_channel]","in":"query","required":false,"type":"string"},{"name":"query[apple_badge]","in":"query","required":false,"type":"string"},{"name":"query[apple_alert]","in":"query","required":false,"type":"string"},{"name":"query[apple_sound]","in":"query","required":false,"type":"string"},{"name":"query[apple_expiry]","in":"query","required":false,"type":"string"},{"name":"query[apple_content_available]","in":"query","required":false,"type":"string"},{"name":"query[google_collapseKey]","in":"query","required":false,"type":"string"},{"name":"query[google_delayWhileIdle]","in":"query","required":false,"type":"string"},{"name":"query[google_timeToLive]","in":"query","required":false,"type":"string"},{"name":"query[payload]","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Resource representation successfully returned"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}},"post":{"tags":["message"],"summary":"Adds a new resource using representation provided in body","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Representation of the resource that is going to be updated","required":false,"schema":{"$ref":"#/definitions/message"}}],"responses":{"201":{"description":"Resource was successfully created"},"422":{"description":"Resource could not be created due to business validations were not achieved"},"500":{"description":"Error trying to process request"}}}},"/push_notifications/message/{id}":{"get":{"tags":["message"],"summary":"Get an existing representation of the given resource","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"}],"responses":{"200":{"description":"Resource representation successfully returned"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}},"put":{"tags":["message"],"summary":"Update an existing resource using representation provided in request body","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"},{"in":"body","name":"body","description":"Representation of the resource that is going to be updated","required":false,"schema":{"$ref":"#/definitions/message"}}],"responses":{"204":{"description":"Resource was successfully updated"},"404":{"description":"Requested resource does not exist"},"422":{"description":"Resource could not be created due to business validations were not achieved"},"500":{"description":"Error trying to process request"}}},"delete":{"tags":["message"],"summary":"Deletes an existing representation of a given resource","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"}],"responses":{"204":{"description":"Resource was successfully updated"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}}},"/push_notifications/device":{"get":{"tags":["device"],"summary":"Get collection of all resources","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"scope","in":"query","required":false,"type":"string","description":"Use one of the defined query scopes:  all, count_exact_match, count, exact_match"},{"name":"query[identifier]","in":"query","required":false,"type":"string"},{"name":"query[password]","in":"query","required":false,"type":"string"},{"name":"query[provider_name]","in":"query","required":false,"type":"string"},{"name":"query[channels]","in":"query","required":false,"type":"string"}],"responses":{"200":{"description":"Resource representation successfully returned"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}},"post":{"tags":["device"],"summary":"Adds a new resource using representation provided in body","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Representation of the resource that is going to be updated","required":false,"schema":{"$ref":"#/definitions/device"}}],"responses":{"201":{"description":"Resource was successfully created"},"422":{"description":"Resource could not be created due to business validations were not achieved"},"500":{"description":"Error trying to process request"}}}},"/push_notifications/device/{id}":{"get":{"tags":["device"],"summary":"Get an existing representation of the given resource","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"}],"responses":{"200":{"description":"Resource representation successfully returned"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}},"put":{"tags":["device"],"summary":"Update an existing resource using representation provided in request body","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"},{"in":"body","name":"body","description":"Representation of the resource that is going to be updated","required":false,"schema":{"$ref":"#/definitions/device"}}],"responses":{"204":{"description":"Resource was successfully updated"},"404":{"description":"Requested resource does not exist"},"422":{"description":"Resource could not be created due to business validations were not achieved"},"500":{"description":"Error trying to process request"}}},"delete":{"tags":["device"],"summary":"Deletes an existing representation of a given resource","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"}],"responses":{"204":{"description":"Resource was successfully updated"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}}},"/activity":{"get":{"tags":["activity"],"summary":"Get collection of all resources","consumes":["application/json"],"produces":["application/json"],"parameters":[],"responses":{"200":{"description":"Resource representation successfully returned"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}},"post":{"tags":["activity"],"summary":"Adds a new resource using representation provided in body","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","description":"Representation of the resource that is going to be updated","required":false,"schema":{"$ref":"#/definitions/activity"}}],"responses":{"201":{"description":"Resource was successfully created"},"422":{"description":"Resource could not be created due to business validations were not achieved"},"500":{"description":"Error trying to process request"}}}},"/activity/{id}":{"get":{"tags":["activity"],"summary":"Get an existing representation of the given resource","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"}],"responses":{"200":{"description":"Resource representation successfully returned"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}},"put":{"tags":["activity"],"summary":"Update an existing resource using representation provided in request body","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"},{"in":"body","name":"body","description":"Representation of the resource that is going to be updated","required":false,"schema":{"$ref":"#/definitions/activity"}}],"responses":{"204":{"description":"Resource was successfully updated"},"404":{"description":"Requested resource does not exist"},"422":{"description":"Resource could not be created due to business validations were not achieved"},"500":{"description":"Error trying to process request"}}},"delete":{"tags":["activity"],"summary":"Deletes an existing representation of a given resource","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"Id of the resource to be operated","required":true,"type":"string"}],"responses":{"204":{"description":"Resource was successfully updated"},"404":{"description":"Requested resource does not exist"},"500":{"description":"Error trying to process request"}}}}},"definitions":{"publishes":{"properties":{"id":{"type":"integer"},"channel":{"type":"string"},"message":{"type":"string"},"pubkey":{"type":"string"},"sig":{"type":"string"},"subkey":{"type":"string"},"timestamp":{"type":"string"}}},"channel":{"properties":{}},"message":{"properties":{}},"device":{"properties":{}},"activity":{"properties":{}}}}